extremism_project/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
crispy-bootstrap4==2023.1
cryptography==41.0.5
daphne==3.0.2
distro==1.8.0
Django==4.2.7
django-channels==0.7.0
django-crispy-forms==2.1
ecs-logging==2.1.0
elastic-apm==6.19.0
exceptiongroup==1.1.3
filelock==3.13.1
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
langdetect
logtail-python==0.2.7
lxml==4.9.3
msgpack==1.0.7
multidict==6.0.4
oauthlib==3.2.2
# openai==1.31.1
openai==1.31.1
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic_core==2.14.5
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sentry-sdk==1.34.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
snscrape==0.7.0.20230622
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.14.1
yarl==1.9.3
zope.interface==6.1
django-rosetta==0.10.0
chromadb==0.5.0
groq==0.5.0
langchain==0.2.2
langchain-community==0.2.3
langchain-core==0.2.2
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langcodes==3.3.0
langdetect==1.0.9
langgraph==0.1.1
langsmith==0.1.54
openapi-schema-pydantic==1.2.4
pydantic==1.10.15
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.8.0
unstructured==0.14.4
unstructured[md]
